Move to callbacks, add import try/catch
