+++ keras_cv/callbacks/__init__.py
@@ -1,0 +1,2 @@
+from keras_cv.metrics.coco.pycoco_utils import compute_pycoco_metrics
+from keras_cv.metrics.coco.recall import COCORecall
