+++ keras_cv/metrics/__init__.py
@@ -12,6 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_cv.metrics.coco.coco_pymetric import COCOPyMetric
 from keras_cv.metrics.coco.mean_average_precision import COCOMeanAveragePrecision
 from keras_cv.metrics.coco.recall import COCORecall
