+++ keras_cv/metrics/coco/__init__.py
@@ -14,5 +14,3 @@
 
 from keras_cv.metrics.coco.coco_pymetric import COCOPyMetric
 from keras_cv.metrics.coco.mean_average_precision import COCOMeanAveragePrecision
-from keras_cv.metrics.coco.pycoco_utils import compute_pycoco_metrics
-from keras_cv.metrics.coco.recall import COCORecall
